Batch decimal digit accumulation to block width during parse
AccumulateDecimalDigitsIntoBigInteger batched 9 digits per iteration (the largest power of 10 that fits a uint), so 64-bit builds ran ~2x the multiply/add iterations at half block-width utilization. Scale the batch to the block width -- 19 digits on 64-bit via DigitsToUInt64, 9 on 32-bit -- and widen Add(uint) to Add(nuint) accordingly.

+++ src/libraries/System.Private.CoreLib/src/System/Number.BigInteger.cs
@@ -898,12 +898,12 @@ namespace System
                 SetValue(out result, ref lhs);
             }
 
-            public void Add(uint value)
+            public void Add(nuint value)
             {
                 int length = (int)_length;
                 if (length == 0)
                 {
-                    SetUInt32(out this, value);
+                    SetUInt64(out this, value);
                     return;
                 }
